requirements-simple.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
openai
pillow
httpx